:hammer: step3 ui 비즈니스 로직 분리 완료(edit 제외)
